Move custom hooks to new dir
