+++ pages/blog/edit/[slug].tsx
@@ -6,7 +6,7 @@ import { useState } from 'react';
 import Router from 'next/router';
 import { deletePost } from '@/lib/blog';
 import revalidateChanges from '@/lib/revalidate';
-import { useFetchStatus } from '@/utils/useLoadingIndicator';
+import { useFetchStatus } from '@/hooks/useLoadingIndicator';
 
 import Container from '@/components/Container';
 import BlogStyles from '@/components/BlogStyles';
